Added UserEnrollment model and edited feedback permissions
