+++ lms/models/__init__.py
@@ -1,3 +1,4 @@
 from .event_models import Event, EventInstance
-from .user_models import UserEnrollment, User, EventRole
+from .user_models import User, EventRole
+from .enrollment_models import UserEnrollment
 from .feedback_models import EventInstanceFeedback
