+++ lms/models/user_models.py
@@ -9,18 +9,7 @@ class EventRole(models.IntegerChoices):
     EVENT_ADMIN = 3
     COORDINATOR = 4
     LEAD = 5
-
-
-class UserEnrollment(models.Model):
-    ''' Store administrative information about participant'''
-
-    user = models.ForeignKey("User", on_delete=models.CASCADE, null=True)
-    paymentId = models.CharField(max_length=200, blank=True, null=True)
-    eventInstance = models.ForeignKey("EventInstance", on_delete=models.CASCADE, null=True)
-    paymentPlatform = models.CharField(max_length=200, blank=True, null=True)
-    role = models.IntegerField(choices=EventRole.choices, blank=False, null=True)
     
-
 class User(AbstractUser):
     pass
