+++ lms/serializers/enrollment_serializers.py
@@ -3,7 +3,8 @@ from django.shortcuts import get_object_or_404
 from django.db.models import Q
 from rest_framework.exceptions import ValidationError
 from backend.exception_classes import ModelObjectAlreadyExist
-from lms.models.user_models import User, UserEnrollment
+from lms.models.user_models import User
+from lms.models.enrollment_models import UserEnrollment
 from lms.models.event_models import EventInstance
