+++ lms/views/enrollment_views.py
@@ -1,7 +1,7 @@
 from rest_framework import generics, permissions, mixins
 from lms.views.filters.enrollment_filter import EnrollmentFilter
 from django_filters.rest_framework import DjangoFilterBackend
-from lms.models.user_models import UserEnrollment
+from lms.models.enrollment_models import UserEnrollment
 from lms.serializers.enrollment_serializers import EnrollmentSerializer
